time-util: widen jiffies_to_usec() parameter to unsigned long
The kernel exposes jiffies counters as unsigned long values (via
/proc/stat and friends), which can exceed uint32_t on 64-bit systems
with long uptimes. Widen the parameter from uint32_t to unsigned long
and cast to usec_t before the multiplication to prevent truncation.

+++ src/basic/time-util.h
@@ -182,7 +182,7 @@ int mktime_or_timegm_usec(struct tm *tm, bool utc, usec_t *ret);
 int localtime_or_gmtime_usec(usec_t t, bool utc, struct tm *ret);
 
 uint32_t usec_to_jiffies(usec_t usec);
-usec_t jiffies_to_usec(uint32_t jiffies);
+usec_t jiffies_to_usec(unsigned long jiffies);
 
 bool in_utc_timezone(void);
